+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2022-09-16
+
+### Added
+
+- Template helpers `#windows`, `#linx` and `#darwin` which work like `if`s for the respective os
+- `eval` template helper which evaluates the given string on the shell
+
 ## [0.7.1] - 2022-09-12
 
 ### Fixed
@@ -152,7 +159,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dotfile linking
 - Error handling
 
-[Unreleased]: https://github.com/volllly/rotz/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/volllly/rotz/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/volllly/rotz/releases/tag/v0.8.0
 [0.7.1]: https://github.com/volllly/rotz/releases/tag/v0.7.1
 [0.7.0]: https://github.com/volllly/rotz/releases/tag/v0.7.0
 [0.6.1]: https://github.com/volllly/rotz/releases/tag/v0.6.1